feat(issue-734): share demo symbol guard test utility
