+++ model/application/src/cam_orchestration.rs
@@ -295,25 +295,14 @@ pub fn count_non_cutting_interference_segments(
 
 #[cfg(test)]
 mod tests {
-    use std::fs;
-    use std::path::{Path, PathBuf};
-
     use super::*;
     use cam_core::CuttingDirection;
 
-    fn collect_rs_files(dir: &Path, files: &mut Vec<PathBuf>) {
-        let entries = fs::read_dir(dir).expect("source directory should be readable");
-        for entry in entries {
-            let entry = entry.expect("directory entry should be readable");
-            let path = entry.path();
-            if path.is_dir() {
-                collect_rs_files(&path, files);
-                continue;
-            }
-            if path.extension().and_then(|ext| ext.to_str()) == Some("rs") {
-                files.push(path);
-            }
-        }
+    mod demo_symbol_guard {
+        include!(concat!(
+            env!("CARGO_MANIFEST_DIR"),
+            "/../test_support/demo_symbol_guard.rs"
+        ));
     }
 
     #[test]
@@ -454,28 +443,6 @@ mod tests {
 
     #[test]
     fn test_application_source_does_not_reference_demo_symbols() {
-        let src_root = Path::new(env!("CARGO_MANIFEST_DIR")).join("src");
-        let mut files = Vec::new();
-        collect_rs_files(&src_root, &mut files);
-
-        let forbidden_symbols = [
-            ["CamSimulation", "DemoScenario"].concat(),
-            ["build_demo_", "artifacts_for_cam_simulation"].concat(),
-            ["create_demo_", "snapshot_exports_for_scenario"].concat(),
-            ["create_sample_", "snapshot_exports_for_demo"].concat(),
-            ["cam_", "sim_demo"].concat(),
-        ];
-
-        for file in files {
-            let content = fs::read_to_string(&file).expect("source file should be readable");
-            for symbol in &forbidden_symbols {
-                assert!(
-                    !content.contains(symbol),
-                    "application layer must not reference demo symbol '{}' in {}",
-                    symbol,
-                    file.display()
-                );
-            }
-        }
+        demo_symbol_guard::assert_layer_does_not_reference_demo_symbols("application");
     }
 }
