+++ model/cam_sim/src/tests.rs
@@ -1,6 +1,4 @@
-use std::fs;
 use std::io::Cursor;
-use std::path::Path;
 
 use cam_core::{
     ArtifactHeaderV1, ArtifactKind, ContourLevelPath, CuttingDirection, SegmentType, Tool,
@@ -57,22 +55,11 @@ fn roundtrip_toolpath_via_artifact(toolpath: &ToolPath<f64>) -> ToolPath<f64> {
     read_toolpath
 }
 
-fn collect_rs_files(dir: &Path, files: &mut Vec<std::path::PathBuf>) {
-    assert!(
-        dir.exists(),
-        "source directory does not exist: {}",
-        dir.display()
-    );
-
-    for entry in fs::read_dir(dir).expect("source directory should be readable") {
-        let entry = entry.expect("directory entry should be readable");
-        let path = entry.path();
-        if path.is_dir() {
-            collect_rs_files(&path, files);
-        } else if path.extension().is_some_and(|ext| ext == "rs") {
-            files.push(path);
-        }
-    }
+mod demo_symbol_guard {
+    include!(concat!(
+        env!("CARGO_MANIFEST_DIR"),
+        "/../test_support/demo_symbol_guard.rs"
+    ));
 }
 
 #[test]
@@ -917,28 +904,5 @@ fn test_ball_end_mill_z_offset_removes_material_at_shifted_z() {
 
 #[test]
 fn test_cam_sim_source_does_not_reference_demo_symbols() {
-    let src_root = Path::new(env!("CARGO_MANIFEST_DIR")).join("src");
-    let mut files = Vec::new();
-    collect_rs_files(&src_root, &mut files);
-    files.sort();
-
-    let forbidden_symbols = [
-        ["CamSimulation", "DemoScenario"].concat(),
-        ["build_demo_", "artifacts_for_cam_simulation"].concat(),
-        ["create_demo_", "snapshot_exports_for_scenario"].concat(),
-        ["create_sample_", "snapshot_exports_for_demo"].concat(),
-        ["cam_", "sim_demo"].concat(),
-    ];
-
-    for file in files {
-        let content = fs::read_to_string(&file).expect("source file should be readable");
-        for symbol in &forbidden_symbols {
-            assert!(
-                !content.contains(symbol),
-                "cam_sim layer must not reference demo symbol '{}' in {}",
-                symbol,
-                file.display()
-            );
-        }
-    }
+    demo_symbol_guard::assert_layer_does_not_reference_demo_symbols("cam_sim");
 }
